requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cohere==2.6.1
decorator==4.4.2
Flask==2.2.2
Flask-RESTful==0.3.9
google-api-core==2.10.2
google-auth==2.12.0
google-cloud-core==2.3.2
google-cloud-speech==2.16.2
google-cloud-storage==2.5.0
google-cloud-videointelligence==2.8.3
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
html-to-json==2.0.0
idna==3.4
imageio==2.22.1
imageio-ffmpeg==0.4.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
moviepy==1.0.3
numpy==1.23.4
Pillow==9.2.0
proglog==0.1.10
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.12
urlopen==1.0.0
utils==1.0.1
Werkzeug==2.2.2
xmltodict==0.12.0
xmltojson==2.0.1
